requirements.txt
# Web framework
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.30.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Pydantic / Settings
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Google AI (new SDK — replaces deprecated google-generativeai)
google-genai>=1.0.0

# Pinecone
pinecone>=4.0.0

# Security / Encryption
cryptography>=42.0.0
python-jose[cryptography]>=3.3.0

# HTTP client (for JWKS fetch)
httpx>=0.27.0

# SQL parsing / validation
sqlparse>=0.5.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.9

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0